Review fix #2
