+++ components/brave_wallet/browser/internal/orchard_test_utils.cc
@@ -6,8 +6,21 @@
 #include "brave/components/brave_wallet/browser/internal/orchard_test_utils.h"
 
 #include "base/memory/ptr_util.h"
+#include "brave/components/brave_wallet/browser/zcash/rust/orchard_decoded_blocks_bundle.h"
 
 namespace brave_wallet {
+
+// static
+OrchardBlockScanner::Result OrchardTestUtils::CreateResultForTesting(
+    const OrchardTreeState& tree_state,
+    const std::vector<OrchardCommitment>& commitments) {
+  auto builder = orchard::OrchardTestUtils::CreateTestingBuilder();
+  for (const auto& commitment : commitments) {
+    builder->AddCommitment(commitment);
+  }
+  builder->SetPriorTreeState(tree_state);
+  return OrchardBlockScanner::Result{{}, {}, builder->Complete()};
+}
 
 OrchardTestUtils::OrchardTestUtils() {
   orchard_test_utils_impl_ = orchard::OrchardTestUtils::Create();
