+++ components/brave_wallet/browser/zcash/rust/orchard_decoded_blocks_bundle.h
@@ -18,21 +18,6 @@ namespace brave_wallet::orchard {
 // a set of dicovered Orchard spendable notes.
 class OrchardDecodedBlocksBundle {
  public:
-  // Builder is used in tests to create OrchardDecodedBlocksBundle with mocked
-  // commitments
-  class TestingBuilder {
-   public:
-    TestingBuilder() = default;
-    virtual ~TestingBuilder() = default;
-    virtual void AddCommitment(
-        const ::brave_wallet::OrchardCommitment& commitment) = 0;
-    virtual void SetPriorTreeState(
-        const ::brave_wallet::OrchardTreeState& tree_state) = 0;
-    virtual std::unique_ptr<OrchardDecodedBlocksBundle> Complete() = 0;
-  };
-
-  static std::unique_ptr<TestingBuilder> CreateTestingBuilder();
-
   virtual ~OrchardDecodedBlocksBundle() = default;
   virtual std::optional<std::vector<::brave_wallet::OrchardNote>>
   GetDiscoveredNotes() = 0;
