+++ components/brave_wallet/browser/zcash/zcash_shield_sync_service_unittest.cc
@@ -11,6 +11,7 @@
 
 #include "base/files/scoped_temp_dir.h"
 #include "brave/components/brave_wallet/browser/brave_wallet_prefs.h"
+#include "brave/components/brave_wallet/browser/internal/orchard_test_utils.h"
 #include "brave/components/brave_wallet/browser/zcash/zcash_rpc.h"
 #include "brave/components/brave_wallet/browser/zcash/zcash_test_utils.h"
 #include "brave/components/brave_wallet/browser/zcash/zcash_wallet_service.h"
@@ -171,7 +172,7 @@ class ZCashShieldSyncServiceTest : public testing::Test {
               mojom::CoinType::ZEC, mojom::KeyringId::kZCashMainnet,
               mojom::AccountKind::kDerived, 0);
           OrchardBlockScanner::Result result =
-              OrchardBlockScanner::CreateResultForTesting(
+              OrchardTestUtils::CreateResultForTesting(
                   tree_state, std::vector<OrchardCommitment>());
           for (const auto& block : blocks) {
             // 3 notes in the blockchain
@@ -190,7 +191,7 @@ class ZCashShieldSyncServiceTest : public testing::Test {
             if (block->height == kNu5BlockUpdate + 255) {
               result.found_spends.push_back(OrchardNoteSpend{
                   block->height, GenerateMockNullifier(account_id, 1)});
-            } else if (block->height == 265) {
+            } else if (block->height == kNu5BlockUpdate + 265) {
               result.found_spends.push_back(OrchardNoteSpend{
                   block->height, GenerateMockNullifier(account_id, 2)});
             }
@@ -299,7 +300,7 @@ TEST_F(ZCashShieldSyncServiceTest, ScanBlocks) {
                 mojom::CoinType::ZEC, mojom::KeyringId::kZCashMainnet,
                 mojom::AccountKind::kDerived, 0);
             OrchardBlockScanner::Result result =
-                OrchardBlockScanner::CreateResultForTesting(
+                OrchardTestUtils::CreateResultForTesting(
                     tree_state, std::vector<OrchardCommitment>());
             for (const auto& block : blocks) {
               // 3 notes in the blockchain
@@ -371,7 +372,7 @@ TEST_F(ZCashShieldSyncServiceTest, ScanBlocks) {
                  base::expected<OrchardBlockScanner::Result,
                                 OrchardBlockScanner::ErrorCode>)> callback) {
             OrchardBlockScanner::Result result =
-                OrchardBlockScanner::CreateResultForTesting(
+                OrchardTestUtils::CreateResultForTesting(
                     tree_state, std::vector<OrchardCommitment>());
             std::move(callback).Run(std::move(result));
           })));
@@ -416,7 +417,7 @@ TEST_F(ZCashShieldSyncServiceTest, ScanBlocks) {
                 mojom::CoinType::ZEC, mojom::KeyringId::kZCashMainnet,
                 mojom::AccountKind::kDerived, 0);
             OrchardBlockScanner::Result result =
-                OrchardBlockScanner::CreateResultForTesting(
+                OrchardTestUtils::CreateResultForTesting(
                     tree_state, std::vector<OrchardCommitment>());
             for (const auto& block : blocks) {
               // First block is the current chain tip - kChainReorgBlockDelta
